Add files via upload
The task of this project was "The area of a rectangle is the rectangle’s length times its width. Write a program that asks 
for the length and width of two rectangles. The program should tell the user which 
rectangle has the greater area, or if the areas are the same."