Add files via upload
The task for this python project was "One foot equals 12 inches. Write a function named feet_to_inches that accepts a 
number of feet as an argument, and returns the number of inches in that many feet. Use the 
function in a program that prompts the user to enter a number of feet and then displays the 
number of inches in that many feet."